requirements.txt
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv~=0.15.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zope.interface==5.1.2
oauthlib~=3.1.0
pyOpenSSL==19.0.0
Flask-Login==0.4.1
cryptography==3.4.6
argon2~=0.1.10
pip~=21.0.1
wheel~=0.34.2
lxml~=4.6.2
setuptools~=47.1.0